Adaptado beneficiarios a nueva base de datos y usando lombok

+++ apaman-backend/src/main/java/com/apaman/apaman_backend/exception/BeneficiarioNotFoundException.java
@@ -1,7 +1,7 @@
 package com.apaman.apaman_backend.exception;
 
 public class BeneficiarioNotFoundException extends RuntimeException{
-    public BeneficiarioNotFoundException(Long id){
+    public BeneficiarioNotFoundException(int id){
         super("No se pudo encontrar el beneficiario con la cedula: " + id);
     }
 }
